docs: fix example Docker command
The README's Docker snippet was passing "megahit" twice:

  docker run ... vout/megahit \
    megahit -1 reads_R1.fq -2 reads_R2.fq -o out

Because the image's ENTRYPOINT already invokes the binary, the
extra "megahit" made CLI parser show the help menu.
Users could not run the containerized assembler "out of the box."

This change drops the second "megahit" so that flags are correctly
forwarded to the default entrypoint:

  docker run ... vout/megahit \
    -1 reads_R1.fq -2 reads_R2.fq -o out

Now the example works as intended and users can launch MEGAHIT
via Docker without confusion.